test/web-platform-tests/to-upstream/html/semantics/forms/the-input-element/disabled-checkbox.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Disabled checkbox event handling</title>
<link rel="author" title="jeffcarp" href="mailto:gcarpenterv@gmail.com">
<link rel=help href="https://html.spec.whatwg.org/#enabling-and-disabling-form-controls:-the-disabled-attribute">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {
  const input = document.createElement("input");
  input.type = "checkbox";
  input.disabled = true;
  document.body.appendChild(input);
  const events = [];

  input.addEventListener("change", () => {
    events.push("change");
  });
  input.addEventListener("click", () => {
    events.push("click");
  });
  input.addEventListener("input", () => {
    events.push("input");
  });

  // disabled checkbox begins unchecked
  assert_false(input.checked);

  input.click();

  // disabled checkbox still unchecked after .click()
  assert_false(input.checked);
  // disabled checkbox emits no events after synthetic .click()
  assert_array_equals(events, []);

  input.dispatchEvent(new MouseEvent("click", { bubbles: true, cancelable: true }));

  // disabled checkbox is checked after dispatching direct click event
  assert_true(input.checked);
  // disabled checkbox runs click activation events after dispatchEvent
  assert_array_equals(events, ["click", "input", "change"]);

}, "checkbox input respects disabled attribute for synthetic events but not for dispatchEvent");
</script>
